feat: a chat save tag as a startup flag
This branch and workspace is to develop a feature to allow users to pass a chat save tag as a startup flag so that the CLI will begin in either interactive mode or non interactive mode with the specified chat history (based on tag) restored.

You should support this as part of startup whether or not a --prompt argument is also included.

The feature support a command line argument that takes either a tag, or a full path to a chat history json file